+++ index.js
@@ -45,6 +45,7 @@ var CleverTap = {
     CleverTapInboxDidInitialize: CleverTapReact.CleverTapInboxDidInitialize,
     CleverTapInboxMessagesDidUpdate: CleverTapReact.CleverTapInboxMessagesDidUpdate,
     CleverTapInboxMessageButtonTapped: CleverTapReact.CleverTapInboxMessageButtonTapped,
+    CleverTapInboxMessageTapped:CleverTapReact.CleverTapInboxMessageTapped,
     CleverTapDisplayUnitsLoaded: CleverTapReact.CleverTapDisplayUnitsLoaded,
     CleverTapInAppNotificationButtonTapped: CleverTapReact.CleverTapInAppNotificationButtonTapped,
     CleverTapFeatureFlagsDidUpdate: CleverTapReact.CleverTapFeatureFlagsDidUpdate,
@@ -104,12 +105,19 @@ var CleverTap = {
     },
 
     /**
-    * Manually set the push token on the CleverTap user profile
-    * @param {string} token - the device token
-    * @param {string} type - for Android only, specifying the type of push service token. Values can be CleverTap.FCM for Firebase or CleverTap.XPS for Xiaomi or CleverTap.BPS for Baidu or CleverTap.HPS for Huawei
-    */
-    setPushToken: function (token, type) {
-        CleverTapReact.setPushTokenAsString(token, type);
+     * Manually set the push token on the CleverTap user profile
+     * @param {string} token - the device token
+     * @param {string} type - for Android only, specifying the type of push service token. Values can be CleverTap.FCM for Firebase or CleverTap.XPS for Xiaomi or CleverTap.BPS for Baidu or CleverTap.HPS for Huawei,
+     * @param {string} region - for xps only ,to specify the region
+     */
+    setPushToken: function (token, type,region="") {
+        console.log(`CleverTap RN | setPushToken | received : token: '${token}' | type:'${type}' | region:'${region}' `)
+        if(type === CleverTap.XPS){
+            CleverTapReact.setPushTokenAsStringWithRegion(token, type,region);
+        }
+        else {
+            CleverTapReact.setPushTokenAsString(token, type);
+        }
     },
 
     /**